requirements.txt
aiogram==2.25.1
aiohttp>=3.8.0
flask>=2.0.0
gunicorn>=20.0.0
python-dotenv>=1.0.0
